move authSecret export
